+++ examples/using-mongodb/src/layouts/index.js
@@ -20,7 +20,7 @@ class DefaultLayout extends React.Component {
             Example of using mongoDB as a data source for a Gatsby site
           </h3>
         </Link>
-        {this.props.children()}
+        {this.props.children}
         <hr />
         <p>
           The src for this website is at