+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="2.0.1-beta.1"></a>
-## [2.0.1-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.0...gatsby-source-contentful@2.0.1-beta.1) (2018-06-17)
+<a name="2.0.1-beta.6"></a>
+## [2.0.1-beta.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.5...gatsby-source-contentful@2.0.1-beta.6) (2018-06-23)
 
 **Note:** Version bump only for package gatsby-source-contentful
 
@@ -12,7 +12,44 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<a name="2.0.1-beta.5"></a>
+## [2.0.1-beta.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.4...gatsby-source-contentful@2.0.1-beta.5) (2018-06-21)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+
+
+
+
+<a name="2.0.1-beta.4"></a>
+## [2.0.1-beta.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.3...gatsby-source-contentful@2.0.1-beta.4) (2018-06-20)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+
+
+
+
+<a name="2.0.1-beta.3"></a>
+
+## [2.0.1-beta.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.2...gatsby-source-contentful@2.0.1-beta.3) (2018-06-20)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-beta.2"></a>
+
+## [2.0.1-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.1...gatsby-source-contentful@2.0.1-beta.2) (2018-06-19)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-beta.1"></a>
+
+## [2.0.1-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.0...gatsby-source-contentful@2.0.1-beta.1) (2018-06-17)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
 <a name="2.0.1-beta.0"></a>
+
 ## [2.0.1-beta.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@1.3.54...gatsby-source-contentful@2.0.1-beta.0) (2018-06-17)
 
 **Note:** Version bump only for package gatsby-source-contentful