+++ packages/gatsby-source-medium/CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="2.0.0-beta.1"></a>
-# [2.0.0-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-medium/compare/gatsby-source-medium@2.0.0-beta.0...gatsby-source-medium@2.0.0-beta.1) (2018-06-17)
+<a name="2.0.0-beta.2"></a>
+# [2.0.0-beta.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-medium/compare/gatsby-source-medium@2.0.0-beta.1...gatsby-source-medium@2.0.0-beta.2) (2018-06-20)
 
 **Note:** Version bump only for package gatsby-source-medium
 
@@ -12,7 +12,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<a name="2.0.0-beta.1"></a>
+
+# [2.0.0-beta.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-medium/compare/gatsby-source-medium@2.0.0-beta.0...gatsby-source-medium@2.0.0-beta.1) (2018-06-17)
+
+**Note:** Version bump only for package gatsby-source-medium
+
 <a name="2.0.0-beta.0"></a>
+
 # [2.0.0-beta.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-medium/compare/gatsby-source-medium@1.0.14...gatsby-source-medium@2.0.0-beta.0) (2018-06-17)
 
 **Note:** Version bump only for package gatsby-source-medium