+++ packages/gatsby/CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="2.0.0-beta.1"></a>
-# [2.0.0-beta.1](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.0...gatsby@2.0.0-beta.1) (2018-06-17)
+<a name="2.0.0-beta.9"></a>
+# [2.0.0-beta.9](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.8...gatsby@2.0.0-beta.9) (2018-06-21)
 
 **Note:** Version bump only for package gatsby
 
@@ -12,7 +12,71 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<a name="2.0.0-beta.8"></a>
+# [2.0.0-beta.8](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.7...gatsby@2.0.0-beta.8) (2018-06-21)
+
+**Note:** Version bump only for package gatsby
+
+
+
+
+
+<a name="2.0.0-beta.7"></a>
+# [2.0.0-beta.7](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.6...gatsby@2.0.0-beta.7) (2018-06-21)
+
+**Note:** Version bump only for package gatsby
+
+
+
+
+
+<a name="2.0.0-beta.6"></a>
+# [2.0.0-beta.6](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.5...gatsby@2.0.0-beta.6) (2018-06-21)
+
+**Note:** Version bump only for package gatsby
+
+
+
+
+
+<a name="2.0.0-beta.5"></a>
+# [2.0.0-beta.5](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.4...gatsby@2.0.0-beta.5) (2018-06-21)
+
+**Note:** Version bump only for package gatsby
+
+
+
+
+
+<a name="2.0.0-beta.4"></a>
+# [2.0.0-beta.4](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.3...gatsby@2.0.0-beta.4) (2018-06-20)
+
+**Note:** Version bump only for package gatsby
+
+
+
+
+
+<a name="2.0.0-beta.3"></a>
+
+# [2.0.0-beta.3](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.2...gatsby@2.0.0-beta.3) (2018-06-20)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.2"></a>
+
+# [2.0.0-beta.2](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.1...gatsby@2.0.0-beta.2) (2018-06-19)
+
+**Note:** Version bump only for package gatsby
+
+<a name="2.0.0-beta.1"></a>
+
+# [2.0.0-beta.1](https://github.com/gatsbyjs/gatsby/compare/gatsby@2.0.0-beta.0...gatsby@2.0.0-beta.1) (2018-06-17)
+
+**Note:** Version bump only for package gatsby
+
 <a name="2.0.0-beta.0"></a>
+
 # [2.0.0-beta.0](https://github.com/gatsbyjs/gatsby/compare/gatsby@1.9.272...gatsby@2.0.0-beta.0) (2018-06-17)
 
 **Note:** Version bump only for package gatsby